requirement.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
PyYAML==5.4.1
requests==2.26.0
tk==0.1.0
urllib3==1.26.7
